requirements.txt
# Voice Assistant Dependencies

# Core audio processing
pyaudio>=0.2.11
sounddevice>=0.4.6
pvporcupine>=3.0.0
pygame>=2.5.0

# AI/ML APIs
google-generativeai>=0.3.0
openai>=1.0.0
chromadb>=0.4.0

# Async and utilities
aiofiles>=23.0.0
pydantic>=2.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0

# Development/testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
isort>=5.12.0

# Audio testing dependencies
sounddevice>=0.4.6
numpy>=1.24.0

# Visual feedback for demo
colorama>=0.4.6 